requirements.txt
pyTelegramBotAPI==4.15.4
groq==0.4.2
Flask==3.0.2
python-dotenv==1.0.1